refactor: render 함수 분리
