+++ src/dom/render/renderInitialUI.ts
@@ -1,0 +1,10 @@
+import { renderResultSectionContent } from "./renderResultSectionContent";
+
+export const renderInitialUI = () => {
+  renderResultSectionContent({
+    isLoading: true,
+    isError: false,
+    isLastPage: true,
+    movies: [],
+  });
+};
